refactor code of main view
