+++ lib/core/utils/routes.dart
@@ -1,7 +1,7 @@
 import 'package:fit_hive/features/authentication/presentation/views/login_view.dart';
 import 'package:fit_hive/features/authentication/presentation/views/register_view.dart';
 import 'package:fit_hive/features/goals/presentation/views/goals_view.dart';
-import 'package:fit_hive/features/home/presentation/views/home_view.dart';
+import 'package:fit_hive/features/main_pages/presentation/views/main_view.dart';
 import 'package:fit_hive/features/onboarding/presentation/views/onboarding.dart';
 import 'package:fit_hive/features/register_success/presentation/views/register_success_view.dart';
 import 'package:fit_hive/features/welcome/presentation/views/welcome_view.dart';
